+++ wrappers/aws/aws-sdk-go/aws/aws.go
@@ -7,7 +7,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws/request"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/epsagon/epsagon-go/epsagon"
-	"github.com/epsagon/epsagon-go/internal"
 	"github.com/epsagon/epsagon-go/protocol"
 	"github.com/epsagon/epsagon-go/tracer"
 	"log"
@@ -23,7 +22,7 @@ func WrapSession(s *session.Session, args ...context.Context) *session.Session {
 		request.NamedHandler{
 			Name: "github.com/epsagon/epsagon-go/wrappers/aws/aws-sdk-go/aws/aws.go",
 			Fn: func(r *request.Request) {
-				currentTracer := internal.ExtractTracer(args)
+				currentTracer := epsagon.ExtractTracer(args)
 				completeEventData(r, currentTracer)
 			},
 		})
